refactor(workflows): extract exec tools into lib/exec-tools (C3a)
Phase C3a of the workflows/index.ts split. Pulls the ~280-line block of MCP
exec-tool registrations out of activate() into
registerWorkflowExecTools(ctx).

lib/exec-tools.ts registers all 10 workflow exec tools — list,
get_definition, start, list_instances, get_instance, get_step,
complete_step, plus the migrated get_step / submit_step / check_gates
formatters. activate() now calls registerWorkflowExecTools(ctx).

index.ts sheds the imports that were only feeding those tools
(buildTemplateList, resolveSubWorkflows, formatStepContext, getTask,
updateTask, z) and drops from 755 to 472 lines.

Pure behavior-preserving relocation — tool definitions byte-identical
(including the deferred submit_step error-message classification, which is
relocated as-is, not fixed).

Verified: typecheck + eslint + workflows suite (420/0, incl. the
exec-tools.test.ts callTool coverage) + full suite (5124/0) + binary build
+ isolated-home boot smoke (plugin Ready; /definitions + /instances +
/gates/pending → 200).
